requirements.txt
Django==3.1
djangorestframework==3.11.1
python-dateutil==2.8.1
uWSGI>=2.0.18,<2.1